refactor(layout): move ambient backdrop and canvases into layout
