+++ src/components/layout/ambient-backdrop.tsx
