+++ src/components/layout/spectral-background.tsx
